requirements.txt
pymupdf
python-docx